backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
arxiv==2.1.3
attrs==24.2.0
backoff==2.2.1
boilerpy3==1.0.7
cattrs==24.1.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
construct==2.5.3
cryptography==43.0.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
Events==0.5
exceptiongroup==1.2.2
farm-haystack==1.26.3
fastapi==0.115.0
feedparser==6.0.11
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
haystack==0.42
haystack-ai==2.6.0
haystack-experimental==0.2.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
importlib_metadata==8.5.0
inflect==7.4.0
Jinja2==3.1.4
jiter==0.6.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy-imports==0.3.1
MarkupSafe==3.0.0
monotonic==1.6
more-itertools==10.5.0
networkx==3.2.1
num2words==0.5.13
numpy==1.26.4
openai==1.51.1
packaging==24.1
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pefile==2024.8.26
pillow==10.4.0
platformdirs==4.3.6
posthog==3.7.0
prompthub-py==4.0.0
pycparser==2.22
pydantic==1.10.18
pydantic_core==2.23.4
pymongo==4.10.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ptrace==0.9.9
pytz==2024.2
PyYAML==6.0.2
quantulum3==0.9.2
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-cache==0.9.8
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sseclient-py==1.8.0
starlette==0.38.6
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.15.2
tqdm==4.66.5
transformers==4.39.3
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.2
url-normalize==1.4.3
urllib3==2.2.3
uvicorn==0.31.0
zipp==3.20.2
